SignalGenerator: implemented noise controls
SignalGenerator can add noise on top of generated signals. This feature
makes use of gnuradio noise_source block. The noise_source is added to the
generated signal. The impulse noise type generates impulses that are
a few times bigger than the selected amplitude. The amplitude of the pulses
is scaled by 20(experimental value) to bring them back within range.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>