Oscilloscope: restart flow every n buffers
A bug in gnuradio(we think) causes the flow to take a long time when
stopping and starting after a long period of time due to cleaning up.
By restarting the the flow every n buffers, we're distributing the
cleanup time during the measurement so it's not noticable when restarting
the instruments.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>