DetachedWindowsManager: Manager for all the detached windows
This class will act as an object pool with detached windows. When a tool wants to be detached it will request a window from this manager, when it is attached it returns the window to the manager so that it can be reused. In this way we avoid creating and deleting QMainWindows every time a tool is detached/attached

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>