pg_channel_manager: remove unused decc variable assignment
This has been put into it's own commit, because it returns a decoder
object, but at the same time, it's not clear if there is anything wrong
with removing it (i.e. whether the decoder() method does anything behind
the scenes that would break anything).

Generally, it looks like removing it is ok, but it's in a separate commit
to draw more attention to it.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>