ToolLauncher:Disconnect the thread watcher when a device session is closed.
When the device is connected, the thread watcher is initialized and
connected to the worker Q_SLOT, but not disconnected once the device
gets disconnected.
This causes issues when connecting and disconnecting multiple times in a
single Scopy session, and then using the "Calibrate" button to calibrate.
The calibration process gets called multiple times. Because of sync
issues, it's possible that none of the calibration works.
Then the device is left in an unknown state.

Signed-off-by: Alexandra.Trifan <Alexandra.Trifan@analog.com>