Enabling/disabling phonehome.
Added a new checkbox for phonehome in the preferences menu. The user can enable/disable the automatic checking by setting/unsetting this checkbox.
This choice is also persisted in Preferences.ini.

On the first scopy run the user is also asked if he wants to enable the phonehome mechanism, via a message box. This message box only appears at the first run.
The first_application_run attribute is used to check if the current run is the first scopy run or not.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>