Disable auto screen scaling on Linux
The QT_AUTO_SCREEN_SCALE_FACTOR makes the application aware of the
system scaling value that is set for the display. Unfortunately this
only works on Windows. On Linux setting this environment variable will
make the application way to big for the screen to fit and unresizable.
The solution for this problem is to enable the environment variable only
on Windows systems.

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>