Oscilloscope: workaround for autoset
autoset uses some gnuradio blocks that are added to the acquisition
flow. The autoset statemachine is updated in the singlecapturedone()
callback which is called whenever data reaches the scope_sink, however
there is no guarantee that data was processed in the autoset FFT.

The workaround consists in not tying the singlecapture done callback to
a step in the autoset state machine, but rather verify the sinks whenever
data reaches the sink.

This bug was probably triggered when moving to gnuradion 3.8 with the
new scheduler implementation.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>